requirements.txt
mega.py
colorama
gtts
pyinstaller